Rename GraphicalPointer.h/cpp to WindowPointer.h/cpp
