+++ WindowPointer.h
